MR : Feat: Add Repository of FavoriteMachine, Machine, Question, Vote | JeongYeji => repository

See merge request s09-webmobile3-sub2/S09P12A304!40

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/MachineDto.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class MachineDto {
 
-	private int machineNo;
+	private Long machineNo;
 	private double latitude;
 	private double longitude;
 	private String name;
